Add players to Redis on register

+++ src/models/scores.rs
@@ -86,9 +86,10 @@ pub struct Score {
 
 impl Score {
     /// Calculates and returns the skill points the player earned for this score.
-    pub fn get_skill_points(&self) -> u32 {
+    #[must_use]
+    pub fn get_skill_points(&self) -> i32 {
         let multiplier = (self.league as u32 + 1) * 100;
-        ((self.score as f64 / self.gold_threshold as f64) * multiplier as f64).round() as u32
+        ((f64::from(self.score) / f64::from(self.gold_threshold)) * f64::from(multiplier)).round() as i32
     }
 
     /// Retrieves the scores for a specific song and league, for display in-game.
